forensic-detective/requirements.txt
# Forensic Detective Layer - Python Dependencies

# Image Processing
Pillow==10.2.0
numpy==1.26.3

# Metadata Extraction
piexif==1.1.3

# AWS Integration
boto3==1.34.34
botocore==1.34.34

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
